+++ src/QueryBuilder/AbstractDDLQueryBuilder.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Db\QueryBuilder;
 
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Query\QueryInterface;
-use Yiisoft\Db\Schema\Builder\ColumnInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 use Yiisoft\Db\Schema\SchemaInterface;
 
@@ -41,13 +41,12 @@ abstract class AbstractDDLQueryBuilder implements DDLQueryBuilderInterface
 
     public function addColumn(string $table, string $column, ColumnInterface|string $type): string
     {
-        /** @psalm-suppress DeprecatedMethod */
         return 'ALTER TABLE '
             . $this->quoter->quoteTableName($table)
             . ' ADD '
             . $this->quoter->quoteColumnName($column)
             . ' '
-            . $this->queryBuilder->getColumnType($type);
+            . $this->queryBuilder->buildColumnDefinition($type);
     }
 
     public function addCommentOnColumn(string $table, string $column, string $comment): string
@@ -134,19 +133,15 @@ abstract class AbstractDDLQueryBuilder implements DDLQueryBuilderInterface
             . ' UNIQUE (' . implode(', ', $columns) . ')';
     }
 
-    public function alterColumn(
-        string $table,
-        string $column,
-        ColumnInterface|string $type
-    ): string {
-        /** @psalm-suppress DeprecatedMethod */
+    public function alterColumn(string $table, string $column, ColumnInterface|string $type): string
+    {
         return 'ALTER TABLE '
             . $this->quoter->quoteTableName($table)
             . ' CHANGE '
             . $this->quoter->quoteColumnName($column)
             . ' '
             . $this->quoter->quoteColumnName($column) . ' '
-            . $this->queryBuilder->getColumnType($type);
+            . $this->queryBuilder->buildColumnDefinition($type);
     }
 
     public function checkIntegrity(string $schema = '', string $table = '', bool $check = true): string
@@ -173,13 +168,16 @@ abstract class AbstractDDLQueryBuilder implements DDLQueryBuilderInterface
 
         foreach ($columns as $name => $type) {
             if (is_string($name)) {
-                /** @psalm-suppress DeprecatedMethod */
                 $cols[] = "\t"
                     . $this->quoter->quoteColumnName($name)
                     . ' '
-                    . $this->queryBuilder->getColumnType($type);
+                    . $this->queryBuilder->buildColumnDefinition(
+                        $type instanceof ColumnInterface
+                        ? $type->withName($name)
+                        : $type
+                    );
             } else {
-                /** @psalm-var string $type */
+                /** @var string $type */
                 $cols[] = "\t" . $type;
             }
         }