+++ tests/Provider/ColumnFactoryProvider.php
@@ -7,14 +7,14 @@ namespace Yiisoft\Db\Tests\Provider;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constant\PseudoType;
 use Yiisoft\Db\Expression\Expression;
-use Yiisoft\Db\Schema\Column\BigIntColumnSchema;
-use Yiisoft\Db\Schema\Column\BinaryColumnSchema;
-use Yiisoft\Db\Schema\Column\BooleanColumnSchema;
-use Yiisoft\Db\Schema\Column\DoubleColumnSchema;
-use Yiisoft\Db\Schema\Column\IntegerColumnSchema;
-use Yiisoft\Db\Schema\Column\JsonColumnSchema;
-use Yiisoft\Db\Schema\Column\StringColumnSchema;
-use Yiisoft\Db\Schema\Column\StructuredColumnSchema;
+use Yiisoft\Db\Schema\Column\BigIntColumn;
+use Yiisoft\Db\Schema\Column\BinaryColumn;
+use Yiisoft\Db\Schema\Column\BooleanColumn;
+use Yiisoft\Db\Schema\Column\DoubleColumn;
+use Yiisoft\Db\Schema\Column\IntegerColumn;
+use Yiisoft\Db\Schema\Column\JsonColumn;
+use Yiisoft\Db\Schema\Column\StringColumn;
+use Yiisoft\Db\Schema\Column\StructuredColumn;
 
 class ColumnFactoryProvider
 {
@@ -22,12 +22,12 @@ class ColumnFactoryProvider
     {
         return [
             // definition, expected type, expected instance of, expected column method results
-            '' => ['', ColumnType::STRING, StringColumnSchema::class, ['getDbType' => '']],
-            'text' => ['text', ColumnType::TEXT, StringColumnSchema::class, ['getDbType' => 'text']],
-            'text NOT NULL' => ['text NOT NULL', ColumnType::TEXT, StringColumnSchema::class, ['getDbType' => 'text', 'isNotNull' => true]],
-            'char(1)' => ['char(1)', ColumnType::CHAR, StringColumnSchema::class, ['getDbType' => 'char', 'getSize' => 1]],
-            'decimal(10,2)' => ['decimal(10,2)', ColumnType::DECIMAL, DoubleColumnSchema::class, ['getDbType' => 'decimal', 'getSize' => 10, 'getScale' => 2]],
-            'bigint UNSIGNED' => ['bigint UNSIGNED', ColumnType::BIGINT, BigIntColumnSchema::class, ['getDbType' => 'bigint', 'isUnsigned' => true]],
+            '' => ['', ColumnType::STRING, StringColumn::class, ['getDbType' => '']],
+            'text' => ['text', ColumnType::TEXT, StringColumn::class, ['getDbType' => 'text']],
+            'text NOT NULL' => ['text NOT NULL', ColumnType::TEXT, StringColumn::class, ['getDbType' => 'text', 'isNotNull' => true]],
+            'char(1)' => ['char(1)', ColumnType::CHAR, StringColumn::class, ['getDbType' => 'char', 'getSize' => 1]],
+            'decimal(10,2)' => ['decimal(10,2)', ColumnType::DECIMAL, DoubleColumn::class, ['getDbType' => 'decimal', 'getSize' => 10, 'getScale' => 2]],
+            'bigint UNSIGNED' => ['bigint UNSIGNED', ColumnType::BIGINT, BigIntColumn::class, ['getDbType' => 'bigint', 'isUnsigned' => true]],
         ];
     }
 
@@ -35,12 +35,12 @@ class ColumnFactoryProvider
     {
         return [
             // pseudo-type, expected type, expected instance of, expected column method results
-            'pk' => [PseudoType::PK, ColumnType::INTEGER, IntegerColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
-            'upk' => [PseudoType::UPK, ColumnType::INTEGER, IntegerColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true, 'isUnsigned' => true]],
-            'bigpk' => [PseudoType::BIGPK, ColumnType::BIGINT, IntegerColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
-            'ubigpk' => [PseudoType::UBIGPK, ColumnType::BIGINT, BigIntColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true, 'isUnsigned' => true]],
-            'uuid_pk' => [PseudoType::UUID_PK, ColumnType::UUID, StringColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
-            'uuid_pk_seq' => [PseudoType::UUID_PK_SEQ, ColumnType::UUID, StringColumnSchema::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
+            'pk' => [PseudoType::PK, ColumnType::INTEGER, IntegerColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
+            'upk' => [PseudoType::UPK, ColumnType::INTEGER, IntegerColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true, 'isUnsigned' => true]],
+            'bigpk' => [PseudoType::BIGPK, ColumnType::BIGINT, IntegerColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
+            'ubigpk' => [PseudoType::UBIGPK, ColumnType::BIGINT, BigIntColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true, 'isUnsigned' => true]],
+            'uuid_pk' => [PseudoType::UUID_PK, ColumnType::UUID, StringColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
+            'uuid_pk_seq' => [PseudoType::UUID_PK_SEQ, ColumnType::UUID, StringColumn::class, ['isPrimaryKey' => true, 'isAutoIncrement' => true]],
         ];
     }
 
@@ -48,26 +48,26 @@ class ColumnFactoryProvider
     {
         return [
             // type, expected type, expected instance of
-            'uuid' => [ColumnType::UUID, ColumnType::UUID, StringColumnSchema::class],
-            'char' => [ColumnType::CHAR, ColumnType::CHAR, StringColumnSchema::class],
-            'string' => [ColumnType::STRING, ColumnType::STRING, StringColumnSchema::class],
-            'text' => [ColumnType::TEXT, ColumnType::TEXT, StringColumnSchema::class],
-            'binary' => [ColumnType::BINARY, ColumnType::BINARY, BinaryColumnSchema::class],
-            'boolean' => [ColumnType::BOOLEAN, ColumnType::BOOLEAN, BooleanColumnSchema::class],
-            'tinyint' => [ColumnType::TINYINT, ColumnType::TINYINT, IntegerColumnSchema::class],
-            'smallint' => [ColumnType::SMALLINT, ColumnType::SMALLINT, IntegerColumnSchema::class],
-            'integer' => [ColumnType::INTEGER, ColumnType::INTEGER, IntegerColumnSchema::class],
-            'bigint' => [ColumnType::BIGINT, ColumnType::BIGINT, IntegerColumnSchema::class],
-            'float' => [ColumnType::FLOAT, ColumnType::FLOAT, DoubleColumnSchema::class],
-            'double' => [ColumnType::DOUBLE, ColumnType::DOUBLE, DoubleColumnSchema::class],
-            'decimal' => [ColumnType::DECIMAL, ColumnType::DECIMAL, DoubleColumnSchema::class],
-            'money' => [ColumnType::MONEY, ColumnType::MONEY, StringColumnSchema::class],
-            'datetime' => [ColumnType::DATETIME, ColumnType::DATETIME, StringColumnSchema::class],
-            'timestamp' => [ColumnType::TIMESTAMP, ColumnType::TIMESTAMP, StringColumnSchema::class],
-            'time' => [ColumnType::TIME, ColumnType::TIME, StringColumnSchema::class],
-            'date' => [ColumnType::DATE, ColumnType::DATE, StringColumnSchema::class],
-            'structured' => [ColumnType::STRUCTURED, ColumnType::STRUCTURED, StructuredColumnSchema::class],
-            'json' => [ColumnType::JSON, ColumnType::JSON, JsonColumnSchema::class],
+            'uuid' => [ColumnType::UUID, ColumnType::UUID, StringColumn::class],
+            'char' => [ColumnType::CHAR, ColumnType::CHAR, StringColumn::class],
+            'string' => [ColumnType::STRING, ColumnType::STRING, StringColumn::class],
+            'text' => [ColumnType::TEXT, ColumnType::TEXT, StringColumn::class],
+            'binary' => [ColumnType::BINARY, ColumnType::BINARY, BinaryColumn::class],
+            'boolean' => [ColumnType::BOOLEAN, ColumnType::BOOLEAN, BooleanColumn::class],
+            'tinyint' => [ColumnType::TINYINT, ColumnType::TINYINT, IntegerColumn::class],
+            'smallint' => [ColumnType::SMALLINT, ColumnType::SMALLINT, IntegerColumn::class],
+            'integer' => [ColumnType::INTEGER, ColumnType::INTEGER, IntegerColumn::class],
+            'bigint' => [ColumnType::BIGINT, ColumnType::BIGINT, IntegerColumn::class],
+            'float' => [ColumnType::FLOAT, ColumnType::FLOAT, DoubleColumn::class],
+            'double' => [ColumnType::DOUBLE, ColumnType::DOUBLE, DoubleColumn::class],
+            'decimal' => [ColumnType::DECIMAL, ColumnType::DECIMAL, DoubleColumn::class],
+            'money' => [ColumnType::MONEY, ColumnType::MONEY, StringColumn::class],
+            'datetime' => [ColumnType::DATETIME, ColumnType::DATETIME, StringColumn::class],
+            'timestamp' => [ColumnType::TIMESTAMP, ColumnType::TIMESTAMP, StringColumn::class],
+            'time' => [ColumnType::TIME, ColumnType::TIME, StringColumn::class],
+            'date' => [ColumnType::DATE, ColumnType::DATE, StringColumn::class],
+            'structured' => [ColumnType::STRUCTURED, ColumnType::STRUCTURED, StructuredColumn::class],
+            'json' => [ColumnType::JSON, ColumnType::JSON, JsonColumn::class],
         ];
     }